requirements/default.txt
Django>=1.7,<1.9
click==5.1
colorama==0.3.3
PyYAML==3.11
selenium==2.48.0
